add functions to the correct component files.
