+++ src/components/Notification/Messaging.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Messaging = () => (
+    <div>
+        <h1>*This is your received messages.*</h1>
+    </div>
+);
+
+export default Messaging;
