+++ src/components/Notification/Notification.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Notification = () => (
+    <div>
+      <h1>This is your notifications.</h1>
+    </div>
+);
+
+export default Notification;
