+++ src/components/Profile/EdirProfile.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+
+const EditProfile = () => (
+  <div>
+    <h1>Edit your profile.</h1>
+  </div>
+);
+
+export default EditProfile;
