+++ src/components/Project/CreateProject.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+const CreateProject = () => (
+  <div>
+    <h1>Create a new project.</h1>
+  </div>
+);
+
+export default CreateProject;
