+++ src/components/Project/ProjectsList.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
 const ProjectsList = () => (
-    <div>
-      <h1>Here you can browse list of all projects.</h1>
-    </div>
-  );
+  <div>
+    <h1>Here you can browse list of all projects.</h1>
+  </div>
+);
 
 export default ProjectsList;
